chore: Simplify corner radius rendering on Skia
