--- src/Uno.UI/UI/Xaml/CornerRadius.skia.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using SkiaSharp;
-using Uno.UI.Helpers.WinUI;
-using Windows.Foundation;
-
-namespace Windows.UI.Xaml;
-
-partial struct CornerRadius
-{
-	
-}
